+++ src/main/java/wooteco/subway/dao/SectionDao.java
@@ -3,7 +3,7 @@ package wooteco.subway.dao;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
-import wooteco.subway.domain.Section;
+import wooteco.subway.domain.section.Section;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
