+++ src/main/java/wooteco/subway/domain/section/DeletionStrategy.java
@@ -1,0 +1,10 @@
+package wooteco.subway.domain.section;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface DeletionStrategy {
+    void delete(List<Section> sections, Long lineId, Long stationId);
+
+    Optional<Section> fixDisconnectedSection(List<Section> sections, Long lineId, Long stationId);
+}
