+++ src/main/java/wooteco/subway/domain/section/Section.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.section;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
